!? Add labeled for Encoders, ready to try ...

+++ algorithms/EisnerO2sib.h
@@ -46,6 +46,6 @@ inline int get_index2_o2sib(int len,int h,int c,int m,int l=0,int ln=1)
 
 extern vector<int>* decodeProjective_o2sib(int length,double* scores);
 extern double* encodeMarginals_o2sib(const int length,const double* scores);
-extern double* LencodeMarginals_o2sib(const int length,const double* scores,const int ln);	//labeled
+extern double* LencodeMarginals_o2sib(const long length,const double* scores,const int ln);	//labeled
 
 #endif /* PARSING_V1_ALGORITHMS_EISNERO2SIB_H_ */
